+++ src/extension.ts
@@ -15,7 +15,8 @@ import {
   execAsTask,
   workspaceRelative,
   extensionConfigurationSet,
-  getTargetName
+  getTargetName,
+  genEnvFile
 } from "./utils";
 import {
   getMesonTargets,
@@ -82,7 +83,11 @@ export async function activate(ctx: vscode.ExtensionContext) {
   controller.createRunProfile("Meson debug test", vscode.TestRunProfileKind.Debug, (request, token) => testDebugHandler(controller, request, token), true)
   controller.createRunProfile("Meson run test", vscode.TestRunProfileKind.Run, (request, token) => testRunHandler(controller, request, token), true)
 
-  let changeHandler = async () => { explorer.refresh(); await rebuildTests(controller); };
+  let changeHandler = async () => {
+    explorer.refresh();
+    await rebuildTests(controller);
+    await genEnvFile(buildDir);
+  };
 
   watcher.onDidChange(changeHandler);
   watcher.onDidCreate(changeHandler);